Update GUI code to use FXML
