--- src/main/java/athena/Main.java
@@ -1,195 +1,0 @@
-package athena;
-
-import java.io.IOException;
-
-import athena.commands.Command;
-import athena.commands.ShutdownCommand;
-import athena.exceptions.InputException;
-import athena.parser.Parser;
-import athena.storage.Storage;
-import athena.tasks.TaskList;
-import athena.ui.DialogBox;
-import athena.ui.Ui;
-import javafx.application.Application;
-import javafx.scene.Scene;
-import javafx.scene.control.Button;
-import javafx.scene.control.Label;
-import javafx.scene.control.ScrollPane;
-import javafx.scene.control.TextField;
-import javafx.scene.image.Image;
-import javafx.scene.image.ImageView;
-import javafx.scene.layout.AnchorPane;
-import javafx.scene.layout.Region;
-import javafx.scene.layout.VBox;
-import javafx.stage.Stage;
-
-/**
- * Provides the point of entry into the Athena program. Initializes the program and handles
- * the main program logic.
- */
-public class Main extends Application {
-    private static final String SAVE_DIRECTORY = "data";
-    private static final String SAVE_FILENAME = "athena.txt";
-
-    private TaskList taskList;
-    private final Storage storage;
-    private final Ui ui;
-    private boolean isActive;
-
-    private ScrollPane scrollPane;
-    private VBox dialogContainer;
-    private TextField userInput;
-    private Button sendButton;
-    private Scene scene;
-    private Image user = new Image(this.getClass().getResourceAsStream("/images/DaUser.png"));
-    private Image duke = new Image(this.getClass().getResourceAsStream("/images/DaDuke.png"));
-
-    /**
-     * Initializes the program by loading task data from disk if any and initializing all
-     * requisite objects and data structures.
-     */
-    public Main() {
-        storage = new Storage(SAVE_DIRECTORY, SAVE_FILENAME);
-        initTaskList(); // Load save data if present
-        ui = new Ui(taskList);
-        isActive = true;
-        ui.sayText("Greetings! My name is Athena. What can I help you with?");
-    }
-
-    private void initTaskList() {
-        if (storage.hasExistingSave()) {
-            try {
-                taskList = storage.loadFromDisk();
-            } catch (IOException e) {
-                ui.sayText("I couldn't load from disk. Opening new task list instead.");
-                taskList = new TaskList();
-            }
-        } else {
-            taskList = new TaskList();
-        }
-    }
-
-    /**
-     * Starts running the main logic of the program, which is to keep reading and running user
-     * commands until the 'bye' command is given. Also, saves the task list to the disk
-     * when modified.
-     */
-    public void run() {
-        while (isActive) {
-            String input = ui.readNextLine();
-            try {
-                Command command = Parser.getCommand(input);
-                command.execute(ui, taskList);
-                if (command instanceof ShutdownCommand) {
-                    isActive = false;
-                }
-            } catch (InputException e) {
-                ui.sayText(e.getMessage());
-            }
-            saveIfTaskListModified();
-        }
-    }
-
-    private void saveIfTaskListModified() {
-        if (taskList.wasModified()) {
-            try {
-                storage.saveToDisk(taskList);
-                taskList.setNotModified();
-            } catch (IOException e) {
-                ui.sayText("I encountered a problem saving to disk: " + e.getMessage());
-            }
-        }
-    }
-
-    @Override
-    public void start(Stage stage) {
-        //Step 1. Setting up required components
-
-        //The container for the content of the chat to scroll.
-        scrollPane = new ScrollPane();
-        dialogContainer = new VBox();
-        scrollPane.setContent(dialogContainer);
-
-        userInput = new TextField();
-        sendButton = new Button("Send");
-
-        AnchorPane mainLayout = new AnchorPane();
-        mainLayout.getChildren().addAll(scrollPane, userInput, sendButton);
-
-        scene = new Scene(mainLayout);
-
-        stage.setScene(scene);
-
-        // Step 2. Formatting the window to look as expected
-        stage.setTitle("Athena");
-        stage.setResizable(false);
-        stage.setMinHeight(600.0);
-        stage.setMinWidth(400.0);
-
-        mainLayout.setPrefSize(400.0, 600.0);
-
-        scrollPane.setPrefSize(385, 535);
-        scrollPane.setHbarPolicy(ScrollPane.ScrollBarPolicy.NEVER);
-        scrollPane.setVbarPolicy(ScrollPane.ScrollBarPolicy.ALWAYS);
-
-        scrollPane.setVvalue(1.0);
-        scrollPane.setFitToWidth(true);
-
-        // You will need to import `javafx.scene.layout.Region` for this.
-        dialogContainer.setPrefHeight(Region.USE_COMPUTED_SIZE);
-
-        userInput.setPrefWidth(325.0);
-
-        sendButton.setPrefWidth(55.0);
-
-        AnchorPane.setTopAnchor(scrollPane, 1.0);
-
-        AnchorPane.setBottomAnchor(sendButton, 1.0);
-        AnchorPane.setRightAnchor(sendButton, 1.0);
-
-        AnchorPane.setLeftAnchor(userInput , 1.0);
-        AnchorPane.setBottomAnchor(userInput, 1.0);
-
-        //Scroll down to the end every time dialogContainer's height changes.
-        dialogContainer.heightProperty().addListener((observable) -> scrollPane.setVvalue(1.0));
-
-        //Part 3. Add functionality to handle user input.
-        sendButton.setOnMouseClicked((event) -> {
-            handleUserInput();
-        });
-
-        userInput.setOnAction((event) -> {
-            handleUserInput();
-        });
-
-        stage.show();
-
-    }
-
-    /**
-     * Iteration 2:
-     * Creates two dialog boxes, one echoing user input and the other containing Duke's reply and then appends them to
-     * the dialog container. Clears the user input after processing.
-     */
-    private void handleUserInput() {
-        Label userText = new Label(userInput.getText());
-        Label dukeText = new Label(getResponse(userInput.getText()));
-        dialogContainer.getChildren().addAll(
-                DialogBox.getUserDialog(userText, new ImageView(user)),
-                DialogBox.getDukeDialog(dukeText, new ImageView(duke))
-        );
-        userInput.clear();
-    }
-
-    /**
-     * You should have your own function to generate a response to user input.
-     * Replace this stub with your completed method.
-     */
-    private String getResponse(String input) {
-        return "Duke heard: " + input;
-    }
-
-    public static void main(String[] args) {
-        new Main().run();
-    }
-}
